refactor: PlaylistCarousel 위젯 분리 (#217)
