--- src/pages/curation/ui/index.ts
@@ -1,1 +1,0 @@
-export { default as CurationCarousel } from './CurationCarousel'
