+++ src/pages/feed/ui/index.ts
@@ -1,7 +1,6 @@
 export { default as FollowTab } from './FollowTab'
 export { default as FeedCdList } from './FeedCdList'
 export { default as FeedProfile } from './FeedProfile'
-export { default as CarouselItem } from './CarouselItem'
 export { default as FollowLayout } from './layout/FollowLayout'
 export { default as FeedCdListLayout } from './layout/FeedCdListLayout'
 export { default as FeedCarousel } from './FeedCarousel'
