+++ src/widgets/playlist/index.ts
@@ -10,3 +10,4 @@ export { default as PlaylistInfo } from './PlaylistInfo'
 export { default as VolumeButton } from './VolumeButton'
 export { default as ControlBar } from './ControlBar'
 export { default as PlayerLayout } from './PlayerLayout'
+export { default as PlaylistCarousel } from './PlaylistCarousel'
